Enhanced a lot of testing routines
First release can be prepared.

+++ test.sh
@@ -33,6 +33,16 @@ if [ -z "$src" ] || [ -z "${main_dir}" ] || [ -z "${destination}" ] || [ "${#set
     exit
 fi
 
+if [ -z "${from_scratch}" ]; then
+    from_scratch=false
+fi
+
+if ${from_scratch}; then
+    if [ -d "${main_dir}" ]; then
+        echo "Remove existing ${main_dir}"
+        rm -r ${main_dir}
+    fi
+fi
 mkdir -p ${main_dir}
 
 if [ -d $src ]; then
@@ -59,11 +69,53 @@ fi
 echo "Register the configured destination"
 echo "${destination}" > DESTINATIONS
 
-for keyword in `awk '{if(NR==1){print $1}}' ./DESTINATIONS` ; do
+run_targets=()
+sync_targets=()
 
-    source ./local_scripts/identify_target.sh ${keyword}
+for keyword in  `awk '{print $1}' ./DESTINATIONS` ; do
+    if [ "${keyword##*_}" == "sync" ]; then
+        sync_targets=(${sync_targets[@]} $keyword)
+        continue
+    fi
 
-    echo ${target} ${dest} ${dest_folder} ${user_at_dest}
+    run_targets=(${run_targets[@]} $keyword)
+
+done
+
+echo "$setups" > SETUPS
+
+echo ""
+echo "##   Testing with following targets/setups   ##"
+echo "###############################################"
+echo ""
+
+for run_target in "${run_targets[@]}"; do
+    for setup in `awk '{print $1}' SETUPS`; do
+        if [ "${setup%%_*}" != "${run_target%%_*}" ]; then
+            continue
+        fi
+        sync_to=""
+        for sync_target in "${sync_targets[@]}"; do
+            if [[ "${sync_target}" =~ "${setup}" ]]; then
+                sync_to="| synchronize to: "${sync_target}
+                break
+            fi
+        done
+        echo "Run on:" $run_target $sync_to "| with setup: "$setup
+    done
+done
+
+
+
+total_runs=0
+for run_target in "${run_targets[@]}"; do
+
+    source ./local_scripts/identify_target.sh ${run_target}
+
+    if ${from_scratch}; then
+        echo "Remove ${dest} if existing..."
+        ssh -t ${user_at_dest} "if [ -d ${dest_folder} ]; then rm -r ${dest_folder}; fi"
+    fi
 
     echo "Register base directory for all setups."
     echo "${target}_base ${dest}/base" >> DESTINATIONS
@@ -76,6 +128,10 @@ for keyword in `awk '{if(NR==1){print $1}}' ./DESTINATIONS` ; do
 
     echo "Run all the given test setups..."
     for setup in `awk '{print $1}' SETUPS`; do
+
+        if [ "${setup%%_*}" != "${run_target%%_*}" ]; then
+            continue
+        fi
 
         echo "  Prepare work directory for setup $setup."
         ssh -t ${user_at_dest} "if [ -d ${dest_folder}/${setup} ]; then rm -r ${dest_folder}/${setup}; fi"
@@ -96,24 +152,71 @@ for keyword in `awk '{if(NR==1){print $1}}' ./DESTINATIONS` ; do
         for d in "${dirs[@]}"; do 
             echo "   with input folder $d"
             ./run.sh ${target}_${setup} prepare-before-run $d
+            let total_runs=total_runs+1
         done
+
+        for sync_target in "${sync_targets[@]}"; do
+            if [[ ! "${sync_target}" =~ "${setup}" ]]; then
+                continue
+            fi
+
+            if ${from_scratch}; then
+                sync_dest="`awk -v dest="${sync_target}" '{if($1==dest){print $2}}' ./DESTINATIONS`"
+                echo "Remove ${sync_dest} if existing..."
+                user_at_sync_dest="${sync_dest%:*}"
+                sync_dest_folder="${sync_dest#*:}"
+                ssh -t ${user_at_sync_dest} "if [ -d ${sync_dest_folder} ]; then rm -r ${sync_dest_folder}; fi"
+            fi
+
+            ./sync.sh "${target}_${setup}" "${sync_target}"
+            break
+        done
+
+         
     done
 
 done
+
+echo "Waiting for test jobs to finish..."
 
 let finished=0
-while  [ ${finished} -lt `cat SETUPS | wc -l` ]; do
-    let finished=0
-    for setup in `awk '{print $1}' SETUPS`; do
-        marker=`ssh -t ${user_at_dest} "if [ -f ${dest_folder}/${setup}/*_finished.txt ]; then echo -n ${setup}; fi"`
-        if [ "$marker" == "$setup" ]; then
-            let finished++
-        fi
+while [ ${finished} -lt ${total_runs} ]; do
+
+    for run_target in "${run_targets[@]}"; do
+
+        source ./local_scripts/identify_target.sh ${run_target} > /dev/null 2>&1
+
+        for setup in `awk '{print $1}' SETUPS`; do
+
+            if [ "${setup%%_*}" != "${run_target%%_*}" ]; then
+                continue
+            fi
+
+            if [ ! -z ${setup_done} ]; then
+                continue
+            fi
+
+            dirs=(`ssh ${user_at_dest} "if [ -d ${dest_folder}/${setup}/input ]; then ls ${dest_folder}/${setup}/input; fi"`)
+            if [[ "${dirs[*]}" =~ "global_settings.py" ]]; then 
+                dirs=("")
+            fi
+            success=`ssh -t ${user_at_dest} "ls ${dest_folder}/${setup}/*_finished.txt 2> /dev/null | wc -l" | sed s.'\r'.''.g`
+            failed=`ssh -t ${user_at_dest} "ls ${dest_folder}/${setup}/*_failed.txt 2> /dev/null | wc -l" | sed s.'\r'.''.g`
+            let completed=0
+            let completed=completed+success+failed
+            if [[ "$completed" == "${#dirs[@]}" ]]; then 
+                let finished=finished+completed
+                eval ${setup}_done=true
+                echo "   Setup $setup done."
+                continue 3 # skip waiting in while loop since this might be the last setup
+            fi
+        done
     done
-    sleep 10
+    sleep 30
 done
+echo "... done."
 
-exit
-
+echo "Create test report..."
 cd -
 ./test_report.sh ${test_config}
+echo "... done."
